rename transactions repository file
